Homing Torpedo and PowerUp Spawner
Homing Torpedo added (requires quick playerShoot change to check if the playerStats have homing or not to choose between standard or homing)

Also added Spawner for power Ups. Comes with two boundaries to determine the X axis spawn, the right boundary is also how high up they will spawn.

FallSpeed and DespawnTimer for powerups are on the powerups themselves